Cancelled Project, HousingControlsBot Source, Wool Sim Code
The recent update gave me the idea of chained together using the push actions. I lost interest due to the fact it didn't work great and would be a pain to implement for all users.

The source on the House for HousingControlsBot from an old video can now be seen in the archive.

Source code to Champlings Wool Sim is now visible.